Add filter to fcm_post_search orderby
Had some complaints from customers that the results when searching for posts in fcm was not satisfactory. Tracked this down to the use of `get_posts` here which sets `orderby => date` as default. This means that ElasticPress will not add it's own `orderby => _score` and the results will be quite bad. To not break BC I've added this as a filter so the sites that uses ElasticPress can do something like this:

```PHP
/**
 * Return an empty string for fcm orderby.
 * This is so that ElasticPress will use _score instead of date for ordering.
 */
add_filter(
	'fcm_post_orderby',
	function(): string {
		return '';
	}
);
```